requirements.txt
flask==3.0.0
flask-cors==4.0.0
unidecode==1.2.0
speedtest-cli==2.1.3